Added date to URL params, devided project into modules
