+++ .gitignore
@@ -1,6 +1,6 @@
 .idea/
 node_modules/
-lesson-15
-coverage
-BOM
-Mastersclass
+lesson-15/
+coverage/
+BOM/
+Mastersclass/
