+++ Mastersclass/Weather.html
@@ -18,7 +18,6 @@
             type="text"
             id="city"
             class="search"
-
             placeholder="Searh your city ..."
           />
           <button class="btn-search" id="search">Search</button>
@@ -47,8 +46,8 @@
             </div>
           </div>
 
-          <div  class="weekly-weather">
-            <div  class="weekly" id="weekly"></div>
+          <div class="weekly-weather">
+            <div class="weekly" id="weekly"></div>
           </div>
         </div>
       </div>
