+++ if-html.html
@@ -63,7 +63,7 @@
               >
               <input
                 class="form__date--in input-style"
-                type="text"
+                type="date"
                 id="date-in"
                 value="Tue 15 Sept"
               />
@@ -75,7 +75,7 @@
               >
               <input
                 class="form__date--out input-style"
-                type="text"
+                type="date"
                 id="date-out"
                 value="Sat 19 Sept"
               />
@@ -518,12 +518,11 @@
         </p>
       </footer>
     </div>
-    <!--    <script src="./node_modules/swiper/swiper-bundle.js"></script>-->
-    <!--    <script src="swiper.js "></script>-->
+<!--        <script src="./node_modules/swiper/swiper-bundle.js"></script>-->
+<!--        <script src="swiper.js "></script>-->
     <script type="module" src="src/scripts/main.js"></script>
     <script type="module" src="src/scripts/module.js"></script>
-    <script type="module" src="src/scripts/functions.js"></script>
+    <script type="module" src="src/scripts/rendering.js"></script>
 
-    <!--    <script type="module" src="src/scripts/Lessons.js"></script>-->
   </body>
 </html>
