+++ src/css/main-styles.css
@@ -664,6 +664,7 @@ figcaption {
   transform: translate(-50%, -50%);
 }
 h5 {
+  text-align: center;
   font-weight: 500;
   font-size: 24px;
   color: var(--main-text);
